Moved shuriken code into shuriken.lua
